requirements-test.txt
-r requirements.txt

coverage
flake8
model-mommy
tox
pytest
pytest-django
